rename some files to follow camelCasing
