+++ packages/browser-sdk/src/flags/maskedProxy.ts
